fixers/vuln-auto-patcher/requirements.txt
langchain
openai
gitpython
pygithub
rich
pytest
python-dotenv
semantic-version
